Fixed rba basic testsuite when both modes are used.
